requirements.txt
#ford.py dependencies

Flask>=0.9
Flask-FlatPages>=0.3
Frozen-Flask>=0.9
PyYAML>=3.10
Pygments==1.6rc1